Rename terminal entrypoint and allow overriding command
