+++ pytest.ini
@@ -7,7 +7,7 @@ addopts =
     --strict-markers
     -v
     --tb=short
-    --cov=victoria_entrypoint
+    --cov=victoria_terminal
     --cov-report=term-missing
 
 # Timeout configuration to prevent hanging tests
